requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
boto3==1.29.0
python-dotenv==1.0.0
uvicorn==0.24.0
pydantic==2.4.2
pytest==7.4.3
httpx==0.25.1
moto[s3,textract]==4.2.7
pytest-asyncio==0.21.1 